Mistake spotted in the previous versions.
When the rank argument was entered as "best" or "worst", it was modified into the numeric equivalent.

When that line was moved into the loop to reflect the fact that "worst" means "worst in the state", the problem was made worse: the value "worst" was change to the numeric value for the first state and then stayed the same for all the other loops. I didn't find the bug until I printed the values of 'num' and 'nrObs' for each iteration of the loop.

In this version, I introduce the variable 'atRank', a copy of the 'num', and I then then go on to modify the copy. The variable 'num' is now copied at each iteration.

Also, to use a loop to enter data in a vector, an empty 1x1 vector needs to be created before the loop.